requirements.txt
python-cfonts
termcolor
paramiko
